requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.57.0
diffusers>=0.24.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Firebase
firebase-admin>=6.2.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# ML/AI utilities
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Evaluation metrics
clip-anytorch>=2.6.0

# Cultural metric dependencies (from metric/cultural_metric)
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
pypdf>=4.2.0
optimum>=1.19.0

# Data handling
pandas>=2.0.0
datasets>=2.14.0

# Logging and config
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Quantization
bitsandbytes>=0.41.0

# Progress bars
tqdm>=4.65.0

# Optional: Specific model dependencies (uncomment as needed)
# FLUX
# git+https://github.com/black-forest-labs/flux

# Qwen-VL (for Qwen3-VL-8B)
qwen-vl-utils>=0.0.8

# Development (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
